organize better codes, using header files
