+++ README.md
@@ -43,3 +43,9 @@ g++ main.o main_test.o -lgtest -lgtest_main -lpthread -o test_executable
 
 ./test_executable
 ```
+
+
+
+# Run main.cpp
+
+g++ main.cpp get_data.cpp -o main.exe -lcurl
